create bundle request
