+++ src/awala.spec.ts
@@ -1,143 +1,10 @@
-import { getModelForClass, type ReturnModelType } from '@typegoose/typegoose';
-import type { Connection, HydratedDocument } from 'mongoose';
 import { jest } from '@jest/globals';
-import { addSeconds } from 'date-fns';
 
-import { setUpTestDbConnection } from './testUtils/db.js';
-import { makeMockLogging, partialPinoLog } from './testUtils/logging.js';
-import {
-  AWALA_PDA,
-  MEMBER_MONGO_ID,
-  MEMBER_PUBLIC_KEY_MONGO_ID,
-  SIGNATURE,
-  TEST_SERVICE_OID,
-} from './testUtils/stubs.js';
-import type { ServiceOptions } from './serviceTypes.js';
-import { MemberBundleRequestModelSchema } from './models/MemberBundleRequest.model.js';
-import { createMemberBundleRequest, postToAwala } from './awala.js';
-import type { MemberBundleRequest } from './schemas/awala.schema.js';
+import { AWALA_PDA } from './testUtils/stubs.js';
+import { postToAwala } from './awala.js';
 import { mockSpy } from './testUtils/jest.js';
 import { AWALA_MIDDLEWARE_ENDPOINT } from './testUtils/eventing/stubs.js';
 import { requireFailureResult } from './testUtils/result.js';
-import { generateKeyPair } from './testUtils/webcrypto.js';
-import { derSerialisePublicKey } from './utilities/webcrypto.js';
-import { MemberPublicKeyModelSchema } from './models/MemberPublicKey.model.js';
-
-const { publicKey } = await generateKeyPair();
-const publicKeyBuffer = await derSerialisePublicKey(publicKey);
-
-describe('createMemberBundleRequest', () => {
-  const getConnection = setUpTestDbConnection();
-  const mockLogging = makeMockLogging();
-  let connection: Connection;
-  let serviceOptions: ServiceOptions;
-  let memberBundleRequestModel: ReturnModelType<typeof MemberBundleRequestModelSchema>;
-  let memberPublicKey: HydratedDocument<MemberPublicKeyModelSchema>;
-  let futureTimestamp: string;
-  let methodInput: MemberBundleRequest;
-  beforeEach(async () => {
-    connection = getConnection();
-    serviceOptions = {
-      dbConnection: connection,
-      logger: mockLogging.logger,
-    };
-    memberBundleRequestModel = getModelForClass(MemberBundleRequestModelSchema, {
-      existingConnection: connection,
-    });
-
-    const memberPublicKeyModel = getModelForClass(MemberPublicKeyModelSchema, {
-      existingConnection: connection,
-    });
-
-    memberPublicKey = await memberPublicKeyModel.create({
-      memberId: MEMBER_MONGO_ID,
-      serviceOid: TEST_SERVICE_OID,
-      publicKey: publicKeyBuffer,
-    });
-
-    futureTimestamp = addSeconds(new Date(), 5).toISOString();
-    methodInput = {
-      publicKeyId: memberPublicKey._id.toString(),
-      memberBundleStartDate: futureTimestamp,
-      awalaPda: AWALA_PDA,
-      signature: SIGNATURE,
-    };
-  });
-
-  test('Valid data should be accepted and stored', async () => {
-    const result = await createMemberBundleRequest(methodInput, serviceOptions);
-
-    const dbResult = await memberBundleRequestModel.findOne({
-      publicKeyId: memberPublicKey._id.toString(),
-    });
-    expect(result).toStrictEqual({
-      didSucceed: true,
-    });
-    expect(dbResult).not.toBeNull();
-    expect(dbResult!.memberId).toBe(MEMBER_MONGO_ID);
-    expect(dbResult!.awalaPda.toString('base64')).toBe(AWALA_PDA);
-    expect(dbResult!.signature.toString('base64')).toBe(SIGNATURE);
-    expect(dbResult!.memberBundleStartDate).toBeDate();
-    expect(dbResult!.memberBundleStartDate.toISOString()).toBe(futureTimestamp);
-    expect(mockLogging.logs).toContainEqual(
-      partialPinoLog('info', 'Member bundle request created', {
-        publicKeyId: memberPublicKey._id.toString(),
-      }),
-    );
-  });
-
-  test('Member bundle data should be updated', async () => {
-    const data = {
-      publicKeyId: memberPublicKey._id.toString(),
-      memberBundleStartDate: new Date(),
-      signature: 'test',
-      awalaPda: 'test',
-      memberId: MEMBER_MONGO_ID,
-    };
-    await memberBundleRequestModel.create(data);
-
-    await createMemberBundleRequest(methodInput, serviceOptions);
-    const dbResult = await memberBundleRequestModel.findOne({
-      publicKeyId: memberPublicKey._id.toString(),
-    });
-    expect(dbResult).not.toBeNull();
-    expect(dbResult!.memberId).toBe(MEMBER_MONGO_ID);
-    expect(dbResult!.awalaPda.toString('base64')).toBe(AWALA_PDA);
-    expect(dbResult!.signature.toString('base64')).toBe(SIGNATURE);
-    expect(dbResult!.memberBundleStartDate).toBeDate();
-    expect(dbResult!.memberBundleStartDate.toISOString()).toBe(futureTimestamp);
-  });
-
-  test('Existing data should not create new entry', async () => {
-    await memberBundleRequestModel.create({
-      ...methodInput,
-      memberId: MEMBER_MONGO_ID,
-    });
-
-    await createMemberBundleRequest(methodInput, serviceOptions);
-
-    const countResult = await memberBundleRequestModel.count({
-      publicKeyId: memberPublicKey._id.toString(),
-    });
-    expect(countResult).toBe(1);
-  });
-
-  test('Non existing public key id should be refused', async () => {
-    const result = await createMemberBundleRequest(
-      {
-        ...methodInput,
-        publicKeyId: MEMBER_PUBLIC_KEY_MONGO_ID,
-      },
-      serviceOptions,
-    );
-
-    expect(result.didSucceed).not.toBeTrue();
-    const dbResult = await memberBundleRequestModel.exists({
-      publicKeyId: MEMBER_PUBLIC_KEY_MONGO_ID,
-    });
-    expect(dbResult).toBeNull();
-  });
-});
 
 describe('postToAwala', () => {
   const mockFetch = mockSpy(jest.spyOn(global, 'fetch'));
